fix(ui): fixed hanging UI-bug when changing projection during rotation
